student dashboard is ready to go
