--- client/app/student/setting/page.tsx
@@ -1,891 +1,0 @@
-"use client";
-
-import { useState } from "react";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { Switch } from "@/components/ui/switch";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Textarea } from "@/components/ui/textarea";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-import { Badge } from "@/components/ui/badge";
-import {
-  Settings,
-  User,
-  Bell,
-  Shield,
-  Palette,
-  Camera,
-  Eye,
-  EyeOff,
-  Smartphone,
-  Lock,
-  Download,
-  Trash2,
-} from "lucide-react";
-
-export default function SettingsPage() {
-  const [showPassword, setShowPassword] = useState(false);
-  const [profileData, setProfileData] = useState({
-    firstName: "Maria",
-    lastName: "Papadopoulou",
-    email: "maria.papadopoulou@student.uom.edu.gr",
-    phone: "+30 694 123 4567",
-    bio: "Computer Science student passionate about AI and machine learning.",
-    dateOfBirth: "1999-05-15",
-    address: "123 University Street, Thessaloniki",
-    emergencyContact: "Dimitris Papadopoulos (+30 694 987 6543)",
-  });
-
-  const [notificationSettings, setNotificationSettings] = useState({
-    email: {
-      grades: true,
-      assignments: true,
-      announcements: true,
-      events: false,
-      messages: true,
-      financial: true,
-    },
-    push: {
-      grades: true,
-      assignments: true,
-      announcements: false,
-      events: false,
-      messages: true,
-      financial: false,
-    },
-    sms: {
-      grades: false,
-      assignments: false,
-      announcements: false,
-      events: false,
-      messages: false,
-      financial: true,
-    },
-  });
-
-  const [privacySettings, setPrivacySettings] = useState({
-    profileVisibility: "students",
-    showEmail: false,
-    showPhone: false,
-    allowMessages: true,
-    dataSharing: false,
-    analyticsTracking: true,
-  });
-
-  const [preferences, setPreferences] = useState({
-    theme: "system",
-    language: "en",
-    timezone: "Europe/Athens",
-    dateFormat: "DD/MM/YYYY",
-    currency: "EUR",
-  });
-
-  const [securitySettings, setSecuritySettings] = useState({
-    twoFactorEnabled: false,
-    loginNotifications: true,
-    sessionTimeout: "30",
-  });
-
-  const connectedDevices = [
-    {
-      id: 1,
-      name: "Personal Laptop",
-      device: "Chrome on Windows 11",
-      location: "Thessaloniki, Greece",
-      lastActive: "Active now",
-      current: true,
-    },
-    {
-      id: 2,
-      name: "Mobile Phone",
-      device: "Safari on iOS 17",
-      location: "Thessaloniki, Greece",
-      lastActive: "2 hours ago",
-      current: false,
-    },
-    {
-      id: 3,
-      name: "University Computer",
-      device: "Firefox on Ubuntu",
-      location: "University of Macedonia",
-      lastActive: "1 day ago",
-      current: false,
-    },
-  ];
-
-  const updateProfileData = (field: string, value: string) => {
-    setProfileData((prev) => ({
-      ...prev,
-      [field]: value,
-    }));
-  };
-
-  const updateNotificationSetting = (
-    method: string,
-    type: string,
-    value: boolean
-  ) => {
-    setNotificationSettings((prev) => ({
-      ...prev,
-      [method]: {
-        ...prev[method as keyof typeof prev],
-        [type]: value,
-      },
-    }));
-  };
-
-  const updatePrivacySetting = (setting: string, value: string | boolean) => {
-    setPrivacySettings((prev) => ({
-      ...prev,
-      [setting]: value,
-    }));
-  };
-
-  const updatePreference = (setting: string, value: string) => {
-    setPreferences((prev) => ({
-      ...prev,
-      [setting]: value,
-    }));
-  };
-
-  const updateSecuritySetting = (setting: string, value: string | boolean) => {
-    setSecuritySettings((prev) => ({
-      ...prev,
-      [setting]: value,
-    }));
-  };
-
-  const revokeDevice = (deviceId: number) => {
-    console.log(`Revoking access for device: ${deviceId}`);
-  };
-
-  return (
-    <div className="space-y-6">
-      <div className="flex justify-between items-center">
-        <div>
-          <h1 className="text-2xl font-bold flex items-center gap-2">
-            <Settings className="h-6 w-6" />
-            Settings
-          </h1>
-          <p className="text-muted-foreground">
-            Manage your account settings and preferences
-          </p>
-        </div>
-      </div>
-
-      <Tabs defaultValue="profile" className="space-y-4">
-        <TabsList className="grid w-full grid-cols-6">
-          <TabsTrigger value="profile">Profile</TabsTrigger>
-          <TabsTrigger value="notifications">Notifications</TabsTrigger>
-          <TabsTrigger value="privacy">Privacy</TabsTrigger>
-          <TabsTrigger value="preferences">Preferences</TabsTrigger>
-          <TabsTrigger value="security">Security</TabsTrigger>
-          <TabsTrigger value="data">Data</TabsTrigger>
-        </TabsList>
-
-        <TabsContent value="profile">
-          <div className="space-y-6">
-            <Card>
-              <CardHeader>
-                <CardTitle className="flex items-center gap-2">
-                  <User className="h-5 w-5" />
-                  Profile Information
-                </CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-6">
-                <div className="flex items-center gap-6">
-                  <Avatar className="h-24 w-24">
-                    <AvatarImage src="/placeholder.svg?height=96&width=96" />
-                    <AvatarFallback className="text-lg">MP</AvatarFallback>
-                  </Avatar>
-                  <div className="space-y-2">
-                    <Button variant="outline">
-                      <Camera className="h-4 w-4 mr-2" />
-                      Change Photo
-                    </Button>
-                    <p className="text-sm text-muted-foreground">
-                      JPG, PNG or GIF. Max size 2MB.
-                    </p>
-                  </div>
-                </div>
-
-                <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                  <div className="space-y-2">
-                    <Label htmlFor="firstName">First Name</Label>
-                    <Input
-                      id="firstName"
-                      value={profileData.firstName}
-                      onChange={(e) =>
-                        updateProfileData("firstName", e.target.value)
-                      }
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="lastName">Last Name</Label>
-                    <Input
-                      id="lastName"
-                      value={profileData.lastName}
-                      onChange={(e) =>
-                        updateProfileData("lastName", e.target.value)
-                      }
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="email">Email Address</Label>
-                    <Input
-                      id="email"
-                      type="email"
-                      value={profileData.email}
-                      onChange={(e) =>
-                        updateProfileData("email", e.target.value)
-                      }
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="phone">Phone Number</Label>
-                    <Input
-                      id="phone"
-                      value={profileData.phone}
-                      onChange={(e) =>
-                        updateProfileData("phone", e.target.value)
-                      }
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="dateOfBirth">Date of Birth</Label>
-                    <Input
-                      id="dateOfBirth"
-                      type="date"
-                      value={profileData.dateOfBirth}
-                      onChange={(e) =>
-                        updateProfileData("dateOfBirth", e.target.value)
-                      }
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="emergencyContact">Emergency Contact</Label>
-                    <Input
-                      id="emergencyContact"
-                      value={profileData.emergencyContact}
-                      onChange={(e) =>
-                        updateProfileData("emergencyContact", e.target.value)
-                      }
-                    />
-                  </div>
-                </div>
-
-                <div className="space-y-2">
-                  <Label htmlFor="address">Address</Label>
-                  <Input
-                    id="address"
-                    value={profileData.address}
-                    onChange={(e) =>
-                      updateProfileData("address", e.target.value)
-                    }
-                  />
-                </div>
-
-                <div className="space-y-2">
-                  <Label htmlFor="bio">Bio</Label>
-                  <Textarea
-                    id="bio"
-                    placeholder="Tell us about yourself..."
-                    value={profileData.bio}
-                    onChange={(e) => updateProfileData("bio", e.target.value)}
-                  />
-                </div>
-
-                <Button>Save Profile Changes</Button>
-              </CardContent>
-            </Card>
-          </div>
-        </TabsContent>
-
-        <TabsContent value="notifications">
-          <Card>
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                <Bell className="h-5 w-5" />
-                Notification Preferences
-              </CardTitle>
-            </CardHeader>
-            <CardContent>
-              <div className="space-y-6">
-                <div>
-                  <h3 className="text-lg font-medium mb-4">
-                    Email Notifications
-                  </h3>
-                  <div className="space-y-3">
-                    {Object.entries(notificationSettings.email).map(
-                      ([type, enabled]) => (
-                        <div
-                          key={type}
-                          className="flex items-center justify-between"
-                        >
-                          <div>
-                            <p className="font-medium capitalize">{type}</p>
-                            <p className="text-sm text-muted-foreground">
-                              Receive {type} notifications via email
-                            </p>
-                          </div>
-                          <Switch
-                            checked={enabled}
-                            onCheckedChange={(value) =>
-                              updateNotificationSetting("email", type, value)
-                            }
-                          />
-                        </div>
-                      )
-                    )}
-                  </div>
-                </div>
-
-                <div>
-                  <h3 className="text-lg font-medium mb-4">
-                    Push Notifications
-                  </h3>
-                  <div className="space-y-3">
-                    {Object.entries(notificationSettings.push).map(
-                      ([type, enabled]) => (
-                        <div
-                          key={type}
-                          className="flex items-center justify-between"
-                        >
-                          <div>
-                            <p className="font-medium capitalize">{type}</p>
-                            <p className="text-sm text-muted-foreground">
-                              Receive {type} push notifications
-                            </p>
-                          </div>
-                          <Switch
-                            checked={enabled}
-                            onCheckedChange={(value) =>
-                              updateNotificationSetting("push", type, value)
-                            }
-                          />
-                        </div>
-                      )
-                    )}
-                  </div>
-                </div>
-
-                <div>
-                  <h3 className="text-lg font-medium mb-4">
-                    SMS Notifications
-                  </h3>
-                  <div className="space-y-3">
-                    {Object.entries(notificationSettings.sms).map(
-                      ([type, enabled]) => (
-                        <div
-                          key={type}
-                          className="flex items-center justify-between"
-                        >
-                          <div>
-                            <p className="font-medium capitalize">{type}</p>
-                            <p className="text-sm text-muted-foreground">
-                              Receive {type} notifications via SMS
-                            </p>
-                          </div>
-                          <Switch
-                            checked={enabled}
-                            onCheckedChange={(value) =>
-                              updateNotificationSetting("sms", type, value)
-                            }
-                          />
-                        </div>
-                      )
-                    )}
-                  </div>
-                </div>
-
-                <div className="pt-4 border-t">
-                  <Button>Save Notification Preferences</Button>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-        </TabsContent>
-
-        <TabsContent value="privacy">
-          <Card>
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                <Shield className="h-5 w-5" />
-                Privacy Settings
-              </CardTitle>
-            </CardHeader>
-            <CardContent>
-              <div className="space-y-6">
-                <div className="space-y-4">
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <Label className="text-base">Profile Visibility</Label>
-                      <p className="text-sm text-muted-foreground">
-                        Who can see your profile information
-                      </p>
-                    </div>
-                    <Select
-                      value={privacySettings.profileVisibility}
-                      onValueChange={(value) =>
-                        updatePrivacySetting("profileVisibility", value)
-                      }
-                    >
-                      <SelectTrigger className="w-40">
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="public">Everyone</SelectItem>
-                        <SelectItem value="students">Students Only</SelectItem>
-                        <SelectItem value="faculty">Faculty Only</SelectItem>
-                        <SelectItem value="private">Private</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <Label className="text-base">Show Email Address</Label>
-                      <p className="text-sm text-muted-foreground">
-                        Allow others to see your email address
-                      </p>
-                    </div>
-                    <Switch
-                      checked={privacySettings.showEmail}
-                      onCheckedChange={(value) =>
-                        updatePrivacySetting("showEmail", value)
-                      }
-                    />
-                  </div>
-
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <Label className="text-base">Show Phone Number</Label>
-                      <p className="text-sm text-muted-foreground">
-                        Allow others to see your phone number
-                      </p>
-                    </div>
-                    <Switch
-                      checked={privacySettings.showPhone}
-                      onCheckedChange={(value) =>
-                        updatePrivacySetting("showPhone", value)
-                      }
-                    />
-                  </div>
-
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <Label className="text-base">Allow Messages</Label>
-                      <p className="text-sm text-muted-foreground">
-                        Allow other students to send you messages
-                      </p>
-                    </div>
-                    <Switch
-                      checked={privacySettings.allowMessages}
-                      onCheckedChange={(value) =>
-                        updatePrivacySetting("allowMessages", value)
-                      }
-                    />
-                  </div>
-
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <Label className="text-base">Data Sharing</Label>
-                      <p className="text-sm text-muted-foreground">
-                        Share anonymized data for research purposes
-                      </p>
-                    </div>
-                    <Switch
-                      checked={privacySettings.dataSharing}
-                      onCheckedChange={(value) =>
-                        updatePrivacySetting("dataSharing", value)
-                      }
-                    />
-                  </div>
-
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <Label className="text-base">Analytics Tracking</Label>
-                      <p className="text-sm text-muted-foreground">
-                        Help improve the platform with usage analytics
-                      </p>
-                    </div>
-                    <Switch
-                      checked={privacySettings.analyticsTracking}
-                      onCheckedChange={(value) =>
-                        updatePrivacySetting("analyticsTracking", value)
-                      }
-                    />
-                  </div>
-                </div>
-
-                <div className="pt-4 border-t">
-                  <Button>Save Privacy Settings</Button>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-        </TabsContent>
-
-        <TabsContent value="preferences">
-          <Card>
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                <Palette className="h-5 w-5" />
-                Preferences
-              </CardTitle>
-            </CardHeader>
-            <CardContent>
-              <div className="space-y-6">
-                <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                  <div className="space-y-2">
-                    <Label>Theme</Label>
-                    <Select
-                      value={preferences.theme}
-                      onValueChange={(value) =>
-                        updatePreference("theme", value)
-                      }
-                    >
-                      <SelectTrigger>
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="light">Light</SelectItem>
-                        <SelectItem value="dark">Dark</SelectItem>
-                        <SelectItem value="system">System</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  <div className="space-y-2">
-                    <Label>Language</Label>
-                    <Select
-                      value={preferences.language}
-                      onValueChange={(value) =>
-                        updatePreference("language", value)
-                      }
-                    >
-                      <SelectTrigger>
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="en">English</SelectItem>
-                        <SelectItem value="el">Greek</SelectItem>
-                        <SelectItem value="fr">French</SelectItem>
-                        <SelectItem value="de">German</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  <div className="space-y-2">
-                    <Label>Timezone</Label>
-                    <Select
-                      value={preferences.timezone}
-                      onValueChange={(value) =>
-                        updatePreference("timezone", value)
-                      }
-                    >
-                      <SelectTrigger>
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="Europe/Athens">
-                          Europe/Athens
-                        </SelectItem>
-                        <SelectItem value="Europe/London">
-                          Europe/London
-                        </SelectItem>
-                        <SelectItem value="America/New_York">
-                          America/New_York
-                        </SelectItem>
-                        <SelectItem value="Asia/Tokyo">Asia/Tokyo</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  <div className="space-y-2">
-                    <Label>Date Format</Label>
-                    <Select
-                      value={preferences.dateFormat}
-                      onValueChange={(value) =>
-                        updatePreference("dateFormat", value)
-                      }
-                    >
-                      <SelectTrigger>
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="DD/MM/YYYY">DD/MM/YYYY</SelectItem>
-                        <SelectItem value="MM/DD/YYYY">MM/DD/YYYY</SelectItem>
-                        <SelectItem value="YYYY-MM-DD">YYYY-MM-DD</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  <div className="space-y-2">
-                    <Label>Currency</Label>
-                    <Select
-                      value={preferences.currency}
-                      onValueChange={(value) =>
-                        updatePreference("currency", value)
-                      }
-                    >
-                      <SelectTrigger>
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="EUR">EUR (€)</SelectItem>
-                        <SelectItem value="USD">USD ($)</SelectItem>
-                        <SelectItem value="GBP">GBP (£)</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-                </div>
-
-                <div className="pt-4 border-t">
-                  <Button>Save Preferences</Button>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-        </TabsContent>
-
-        <TabsContent value="security">
-          <div className="space-y-6">
-            <Card>
-              <CardHeader>
-                <CardTitle className="flex items-center gap-2">
-                  <Lock className="h-5 w-5" />
-                  Security Settings
-                </CardTitle>
-              </CardHeader>
-              <CardContent className="space-y-6">
-                <div className="space-y-4">
-                  <div>
-                    <Label htmlFor="currentPassword">Current Password</Label>
-                    <div className="relative">
-                      <Input
-                        id="currentPassword"
-                        type={showPassword ? "text" : "password"}
-                        placeholder="Enter current password"
-                      />
-                      <Button
-                        type="button"
-                        variant="ghost"
-                        size="sm"
-                        className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
-                        onClick={() => setShowPassword(!showPassword)}
-                      >
-                        {showPassword ? (
-                          <EyeOff className="h-4 w-4" />
-                        ) : (
-                          <Eye className="h-4 w-4" />
-                        )}
-                      </Button>
-                    </div>
-                  </div>
-
-                  <div>
-                    <Label htmlFor="newPassword">New Password</Label>
-                    <Input
-                      id="newPassword"
-                      type="password"
-                      placeholder="Enter new password"
-                    />
-                  </div>
-
-                  <div>
-                    <Label htmlFor="confirmPassword">
-                      Confirm New Password
-                    </Label>
-                    <Input
-                      id="confirmPassword"
-                      type="password"
-                      placeholder="Confirm new password"
-                    />
-                  </div>
-
-                  <Button>Change Password</Button>
-                </div>
-
-                <div className="space-y-4 pt-4 border-t">
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <Label className="text-base">
-                        Two-Factor Authentication
-                      </Label>
-                      <p className="text-sm text-muted-foreground">
-                        Add an extra layer of security to your account
-                      </p>
-                    </div>
-                    <div className="flex items-center gap-2">
-                      {securitySettings.twoFactorEnabled && (
-                        <Badge variant="secondary">Enabled</Badge>
-                      )}
-                      <Switch
-                        checked={securitySettings.twoFactorEnabled}
-                        onCheckedChange={(value) =>
-                          updateSecuritySetting("twoFactorEnabled", value)
-                        }
-                      />
-                    </div>
-                  </div>
-
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <Label className="text-base">Login Notifications</Label>
-                      <p className="text-sm text-muted-foreground">
-                        Get notified when someone logs into your account
-                      </p>
-                    </div>
-                    <Switch
-                      checked={securitySettings.loginNotifications}
-                      onCheckedChange={(value) =>
-                        updateSecuritySetting("loginNotifications", value)
-                      }
-                    />
-                  </div>
-
-                  <div className="flex items-center justify-between">
-                    <div>
-                      <Label className="text-base">Session Timeout</Label>
-                      <p className="text-sm text-muted-foreground">
-                        Automatically log out after inactivity
-                      </p>
-                    </div>
-                    <Select
-                      value={securitySettings.sessionTimeout}
-                      onValueChange={(value) =>
-                        updateSecuritySetting("sessionTimeout", value)
-                      }
-                    >
-                      <SelectTrigger className="w-32">
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="15">15 minutes</SelectItem>
-                        <SelectItem value="30">30 minutes</SelectItem>
-                        <SelectItem value="60">1 hour</SelectItem>
-                        <SelectItem value="120">2 hours</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-                </div>
-              </CardContent>
-            </Card>
-
-            <Card>
-              <CardHeader>
-                <CardTitle className="flex items-center gap-2">
-                  <Smartphone className="h-5 w-5" />
-                  Connected Devices
-                </CardTitle>
-              </CardHeader>
-              <CardContent>
-                <div className="space-y-4">
-                  {connectedDevices.map((device) => (
-                    <div
-                      key={device.id}
-                      className="flex items-center justify-between p-4 border rounded-lg"
-                    >
-                      <div className="flex items-center gap-3">
-                        <div className="p-2 bg-muted rounded-lg">
-                          <Smartphone className="h-4 w-4" />
-                        </div>
-                        <div>
-                          <p className="font-medium">{device.name}</p>
-                          <p className="text-sm text-muted-foreground">
-                            {device.device}
-                          </p>
-                          <p className="text-sm text-muted-foreground">
-                            {device.location}
-                          </p>
-                        </div>
-                      </div>
-                      <div className="flex items-center gap-2">
-                        <div className="text-right">
-                          <p className="text-sm">{device.lastActive}</p>
-                          {device.current && (
-                            <Badge variant="secondary" className="text-xs">
-                              Current
-                            </Badge>
-                          )}
-                        </div>
-                        {!device.current && (
-                          <Button
-                            variant="outline"
-                            size="sm"
-                            onClick={() => revokeDevice(device.id)}
-                          >
-                            Revoke
-                          </Button>
-                        )}
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              </CardContent>
-            </Card>
-          </div>
-        </TabsContent>
-
-        <TabsContent value="data">
-          <Card>
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                <Download className="h-5 w-5" />
-                Data Management
-              </CardTitle>
-            </CardHeader>
-            <CardContent>
-              <div className="space-y-6">
-                <div className="space-y-4">
-                  <div>
-                    <h3 className="text-lg font-medium">Export Your Data</h3>
-                    <p className="text-sm text-muted-foreground">
-                      Download a copy of your personal data stored in our system
-                    </p>
-                  </div>
-                  <div className="flex gap-2">
-                    <Button variant="outline">
-                      <Download className="h-4 w-4 mr-2" />
-                      Export Profile Data
-                    </Button>
-                    <Button variant="outline">
-                      <Download className="h-4 w-4 mr-2" />
-                      Export Academic Records
-                    </Button>
-                  </div>
-                </div>
-
-                <div className="space-y-4 pt-4 border-t">
-                  <div>
-                    <h3 className="text-lg font-medium text-red-600">
-                      Delete Account
-                    </h3>
-                    <p className="text-sm text-muted-foreground">
-                      Permanently delete your account and all associated data.
-                      This action cannot be undone.
-                    </p>
-                  </div>
-                  <Button variant="destructive">
-                    <Trash2 className="h-4 w-4 mr-2" />
-                    Delete Account
-                  </Button>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-        </TabsContent>
-      </Tabs>
-    </div>
-  );
-}
